requirement.txt
aiohttp==3.7.3
alembic==1.4.3
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
bokeh==2.2.3
Brotli==1.0.9
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cssselect2==0.3.0
cycler==0.10.0
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.4
django-debug-toolbar==3.2
djangorestframework==3.12.2
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.8.0
Flask-Cors==3.0.9
Flask-Mail==0.9.0
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Toastr==0.5.4
Flask-WeasyPrint==0.6
Flask-WTF==0.14.3
future==0.18.2
html5lib==1.1
idna==2.10
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.7
jupyter-core==4.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
multidict==5.1.0
npm==0.1.1
numpy==1.19.4
optional-django==0.1.0
packaging==20.4
pandas==1.1.4
parso==0.7.1
pdfkit==0.6.1
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.12.0
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
Pygments==2.7.3
pylint==2.5.3
pyparsing==2.4.7
Pyphen==0.9.5
pytest==6.2.1
python-dateutil==2.6.0
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.4
pywin32==300
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.24.0
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
tinycss2==1.0.2
toml==0.10.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.0.31
wcwidth==0.2.5
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
yarl==1.6.3
